# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tchallenger",
        "description": "",
        "contact": {
            "email": "nathalietaffot@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/public/api/participations": {
            "get": {
                "tags": [
                    "Participation"
                ],
                "summary": "Get Participation",
                "description": "Get Participation",
                "operationId": "indexxxxx",
                "responses": {
                    "201": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "affichage de tous les participation."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "400"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur lors du traitement de la demande"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "500"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de connexion"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Participation"
                ],
                "summary": "Register of participation",
                "description": "Register of participation",
                "operationId": "storeeee",
                "requestBody": {
                    "description": "Enregistrement d'une nouvelle participation",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "votre titre",
                                        "type": "string",
                                        "format": "string",
                                        "example": "laporte"
                                    },
                                    "description": {
                                        "description": "votre description",
                                        "type": "string",
                                        "format": "string",
                                        "example": "kdjfkd  kcvng"
                                    },
                                    "likes": {
                                        "description": "nombre de likes",
                                        "type": "integer",
                                        "format": "1",
                                        "example": "1"
                                    },
                                    "vues": {
                                        "description": "nombre de vues",
                                        "type": "integer",
                                        "format": "1",
                                        "example": "1"
                                    },
                                    "shares": {
                                        "description": "nombre de shares",
                                        "type": "integer",
                                        "format": "1",
                                        "example": "1"
                                    },
                                    "comments": {
                                        "description": "votre comments",
                                        "type": "string",
                                        "format": "string",
                                        "example": "comments"
                                    },
                                    "post_id": {
                                        "description": "votre post_id",
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "3"
                                    },
                                    "media_id": {
                                        "description": "votre media_id",
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "participation bien Creer."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "400"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur lors du traitement de la demande"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "500"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de connexion"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [
                            []
                        ]
                    }
                ]
            }
        },
...
    }
}